# public/swagger.json
{
	"components": {
...
		"schemas": {
			"MenuRole": {
				"properties": {
					"id": {
						"type": "string"
					},
					"menu": {
						"$ref": "#/components/schemas/Menu"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"menu",
					"role",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Menu": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"URL": {
						"type": "string"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/MenuRole"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"URL",
					"roles",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"menus": {
						"items": {
							"$ref": "#/components/schemas/MenuRole"
						},
						"type": "array"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"menus",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Zone": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"code",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Bus": {
				"properties": {
					"id": {
						"type": "string"
					},
					"zone": {
						"$ref": "#/components/schemas/Zone"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"zone",
					"name",
					"code",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Operation": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"bus": {
						"$ref": "#/components/schemas/Bus"
					},
					"code": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"bus",
					"code",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isActive": {
						"type": "boolean"
					},
					"auth": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"line": {
						"$ref": "#/components/schemas/Line"
					},
					"SAPCode": {
						"type": "string"
					},
					"SAPUser": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"username",
					"password",
					"email",
					"role",
					"isActive",
					"auth",
					"created",
					"operation",
					"area",
					"line",
					"SAPCode",
					"SAPUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Area": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"name",
					"code",
					"isActive",
					"created",
					"operation",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Line": {
				"properties": {
					"id": {
						"type": "string"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					}
				},
				"required": [
					"id",
					"area",
					"name",
					"SAPCode",
					"isActive",
					"created",
					"operation",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"isActive": {
						"type": "boolean"
					},
					"SAPCode": {
						"type": "string"
					},
					"SAPUser": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"line": {
						"$ref": "#/components/schemas/Line"
					}
				},
				"required": [
					"name",
					"username",
					"password",
					"email",
					"role",
					"isActive",
					"SAPCode",
					"SAPUser",
					"operation",
					"area",
					"line"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAreaPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"password",
					"email",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ProcessType": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Process": {
				"properties": {
					"id": {
						"type": "string"
					},
					"processId": {
						"type": "string"
					},
					"processType": {
						"$ref": "#/components/schemas/ProcessType"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"processId",
					"processType",
					"name",
					"SAPCode",
					"operation",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Priority": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"SAPCode",
					"operation",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Symptom": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"groupCode": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"SAPCode",
					"groupCode",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Subarea": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"code",
					"SAPCode",
					"isActive",
					"area",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISubareaPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"code",
					"area",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Ubication": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"subarea": {
						"$ref": "#/components/schemas/Subarea"
					},
					"isActive": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"code",
					"subarea",
					"isActive",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUbicationPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"subarea": {
						"$ref": "#/components/schemas/Subarea"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"code",
					"subarea",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPriorityPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"SAPCode",
					"operation",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IProcessPayload": {
				"properties": {
					"processType": {
						"$ref": "#/components/schemas/ProcessType"
					},
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"operation": {
						"$ref": "#/components/schemas/Operation"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"processType",
					"name",
					"SAPCode",
					"operation",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ISymptomPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"SAPCode": {
						"type": "string"
					},
					"groupCode": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"SAPCode",
					"groupCode",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "fixmaint",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Richard HC"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "profile",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "take",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserPayload"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"put": {
				"operationId": "UpdateUserStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/areas": {
			"get": {
				"operationId": "GetAreas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Area"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Area"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "operation",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateArea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Area"
								}
							}
						}
					}
				},
				"tags": [
					"Area"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAreaPayload"
							}
						}
					}
				}
			}
		},
		"/areas/{id}": {
			"get": {
				"operationId": "GetArea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Area"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Area"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/subareas": {
			"get": {
				"operationId": "GetSubareas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Subarea"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Subarea"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateSubarea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Subarea"
								}
							}
						}
					}
				},
				"tags": [
					"Subarea"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISubareaPayload"
							}
						}
					}
				}
			}
		},
		"/subareas/{id}": {
			"get": {
				"operationId": "GetSubarea",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Subarea"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Subarea"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ubications": {
			"get": {
				"operationId": "GetUbications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Ubication"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Ubication"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUbication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Ubication"
								}
							}
						}
					}
				},
				"tags": [
					"Ubication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUbicationPayload"
							}
						}
					}
				}
			}
		},
		"/ubications/{id}": {
			"get": {
				"operationId": "GetUbication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Ubication"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Ubication"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/priorities": {
			"get": {
				"operationId": "GetPriorities",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Priority"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Priority"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "operation",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreatePriority",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Priority"
								}
							}
						}
					}
				},
				"tags": [
					"Priority"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPriorityPayload"
							}
						}
					}
				}
			}
		},
		"/priorities/{id}": {
			"get": {
				"operationId": "GetPriority",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Priority"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Priority"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/processes": {
			"get": {
				"operationId": "GetProceses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Process"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Process"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "profile",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateProcess",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Process"
								}
							}
						}
					}
				},
				"tags": [
					"Process"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProcessPayload"
							}
						}
					}
				}
			}
		},
		"/processes/web": {
			"get": {
				"operationId": "GetProcesesWeb",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Process"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Process"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "profile",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/processes/{id}": {
			"get": {
				"operationId": "GetProcess",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Process"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Process"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/symptoms": {
			"get": {
				"operationId": "GetSymptoms",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Symptom"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Symptom"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "groupCode",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateSymptom",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Symptom"
								}
							}
						}
					}
				},
				"tags": [
					"Symptom"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISymptomPayload"
							}
						}
					}
				}
			}
		},
		"/symptoms/{id}": {
			"get": {
				"operationId": "GetSymptom",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Symptom"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Symptom"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}